+++ mod/main.lua
@@ -6,6 +6,7 @@
 --[[
 
 TODO:
+- show multiplayer races as ranked
 - Implement time offsets, show on the first room of each floor
 - Opponent's shadows